+++ app/src/main/java/com/neo/moneytracker/ui/screens/AnalyticsScreen.kt
@@ -12,8 +12,8 @@ import com.neo.moneytracker.ui.components.MonthlyStatsCard
 
 @Composable
 fun AnalyticsScreen(
-    expenseAmount: Double,
-    incomeAmount: Double,
+    expenseAmount: Int,
+    incomeAmount: Int,
     modifier: Modifier = Modifier
 ) {
     Column(
